feat(web): refactor ip comments to use generic representations add and manage documents code
