+++ appeals/web/src/server/appeals/appeal-details/representations/final-comments/accept/accept.controller.js
