+++ appeals/web/src/server/appeals/appeal-details/representations/final-comments/final-comments.router.js
@@ -7,8 +7,8 @@ import {
 	getRepresentationAttachmentsFolder,
 	withSingularRepresentation
 } from './final-comments.middleware.js';
-import addDocumentRouter from '../representations/document-attachments/add-document.router.js';
-import manageDocumentsRouter from '../representations/document-attachments/manage-documents.router.js';
+import addDocumentRouter from '../document-attachments/add-document.router.js';
+import manageDocumentsRouter from '../document-attachments/manage-documents.router.js';
 import acceptFinalCommentsRouter from './accept/accept.router.js';
 
 const router = createRouter({ mergeParams: true });
