+++ appeals/web/src/server/appeals/appeal-details/representations/interested-party-comments/interested-party-comments.router.js
@@ -3,7 +3,7 @@ import { Router as createRouter } from 'express';
 import { validateAppeal } from '../../appeal-details.middleware.js';
 import addIpCommentRouter from './add-ip-comment/add-ip-comment.router.js';
 import editIpCommentRouter from './edit-ip-comment/edit-ip-comment.router.js';
-import addDocumentRouter from './add-document/add-document.router.js';
+import addDocumentRouter from '../document-attachments/add-document.router.js';
 import viewAndReviewIpCommentRouter from './view-and-review/view-and-review.router.js';
 import redactIpCommentRouter from './redact/redact.router.js';
 import * as controller from './interested-party-comments.controller.js';
