+++ appeals/web/src/server/appeals/appeal-details/representations/interested-party-comments/view-and-review/view-and-review.controller.js
@@ -27,7 +27,7 @@ export const renderViewInterestedPartyComment = render(
 
 export const renderReviewInterestedPartyComment = render(
 	reviewInterestedPartyCommentPage,
-	'patterns/change-page.pattern.njk'
+	'patterns/display-page.pattern.njk'
 );
 
 /**
